split log2journal into multiple files
