+++ collectors/log2journal/Makefile.am
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+AUTOMAKE_OPTIONS = subdir-objects
+MAINTAINERCLEANFILES = $(srcdir)/Makefile.in
+
+dist_noinst_DATA = \
+    README.md \
+    $(NULL)
+
+dist_libconfig_DATA = \
+    log2journal.d/nginx-combined.yaml \
+    $(NULL)
